[feat] finalização do projeto

+++ src/services/frases.js
@@ -22,13 +22,35 @@ const listarFrasesPorLicaoService = async (idLicao) => {
 };
 
 const deletarFraseService = async (fraseId) => {
+  const fraseExistente = await Frases.findById(fraseId);
+
+  if (!fraseExistente) {
+    throw new Error("Frase não encontrada");
+  }
+
+  const audioId = fraseExistente.audioCurto;
+  await deletarAudioService(audioId);
+
   await Frases.findByIdAndDelete(fraseId);
-  await deletarAudioService(fraseId);
 
   return { message: "Frase removida com sucesso" };
 };
 
-const atualizarFraseService = async (fraseId, data) => {
+const atualizarFraseService = async (fraseId, data, files) => {
+  const fraseExistente = await Frases.findById(fraseId);
+
+  if (!fraseExistente) {
+    throw new Error("Frase não encontrada");
+  }
+
+  const audioId = fraseExistente.audioCurto;
+
+  if (files) {
+    await deletarAudioService(audioId);
+    const novoAudioId = await salvarAudio(files);
+    data.audioCurto = novoAudioId;
+  }
+
   return Frases.findByIdAndUpdate(fraseId, data, { new: true });
 };
